move login logout test from metadata tests to logout tests and use D2factory
