+++ core/src/androidTest/java/org/hisp/dhis/android/core/MetadataCallRealIntegrationShould.java
@@ -13,7 +13,6 @@ import org.junit.runner.RunWith;
 
 import java.io.IOException;
 
-@RunWith(AndroidJUnit4.class)
 public class MetadataCallRealIntegrationShould extends AbsStoreTestCase {
     /**
      * A quick integration test that is probably flaky, but will help with finding bugs related to
@@ -69,17 +68,5 @@ public class MetadataCallRealIntegrationShould extends AbsStoreTestCase {
         //This way I can make sure that additive (updates) work as well.
         //The changes could be to one of the programs, adding stuff to it.
         // adding a new program..etc.
-    }
-
-    //@Test
-    public void response_successful_on_login_logout_and_login() throws Exception {
-        retrofit2.Response response = null;
-        response = d2.logIn("android", "Android123").call();
-        assertThat(response.isSuccessful()).isTrue();
-
-        d2.logOut().call();
-
-        response = d2.logIn("android", "Android123").call();
-        assertThat(response.isSuccessful()).isTrue();
     }
 }
